refactor(operations): impl Add for refrence types
Instead of strictly following Rust docs, I'm now implementing the ops
traits for references to the primitive types Point and Vector. This
is a pretty nice idea. However, I will refactor even more and directly
implement the `Copy` trait for these types so I don't have to bother
with ownership when using them.